assistant/requirements.txt
# Core application packages
pyyaml

# AI Model packages
openai

# Probe/Integration packages
mysql-connector-python